Add 6-10 star brews. This changes a bit of code around to add support for level 10 brews, also known as Master Brews. The way it works is it checks if a brew has a brand on it, and before ageing or distilling, it extracts that brewery from the lore, and accesses its level. That level is then used as a sort of modifier, where essentially each level is 1 extra star (with a total of 5 levels).
For cooked items though, its a little bit different. Since the items aren't branded when they're taken directly from a cauldron, we have to take a different approach. We basically check the player who's taking the brew, and then do a check to see if they have a brand, if they do, we access the level from there and then proceed as normal!